Fix Windows build: move syscall.Stat_t usage to platform-specific files
syscall.Stat_t is undefined on Windows, causing the package and its tests
to fail to compile. Extract hardlink detection into hardlink_unix.go
(!windows) and a no-op stub hardlink_windows.go, and guard the three
disk-based hardlink tests behind the same !windows build tag.
